LayoutTests/accessibility/mac/AOM-string-properties.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head>
<body>

<div role="combobox" id="autocomplete"></div>
<div role="checkbox" id="checked"></div>
<div role="tab" id="current"></div>
<div role="button" id="hasPopUp"></div>
<div role="textbox" id="invalid"></div>
<div role="button" id="keyShortcuts"></div>
<div role="banner" id="live"></div>
<div role="slider" id="orientation"></div>
<input id="placeholder">
<div role="switch" id="pressed"></div>
<div role="banner" id="relevant"></div>
<div role="button" id="roleDescription"></div>
<div role="columnheader" id="sort"></div>
<div role="slider" id="valueText"></div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests string type Accessibility Object Model properties.");
    if (window.accessibilityController) {
        var node;
        var axNode;

        testAutocomplete();
        testChecked();
        testCurrent();
        testHasPopUp();
        testInvalid();
        testKeyShortcuts();
        testLive();
        testOrientation();
        testPlaceholder();
        testPressed();
        testRelevant();
        testRoleDescription();
        testSort();
        testValueText();
    }

    function testAutocomplete() {
        debug("\nAutocomplete support");
        node = document.getElementById("autocomplete");
        axNode = accessibilityController.accessibleElementById("autocomplete");
        node.accessibleNode.autocomplete = "inline";
        shouldBeEqualToString("node.accessibleNode.autocomplete", "inline");
        shouldBeEqualToString("axNode.stringAttributeValue('AXAutocompleteValue')", "inline");
    }

    function testChecked() {
        debug("\nChecked support");
        node = document.getElementById("checked");
        axNode = accessibilityController.accessibleElementById("checked");
        shouldBeFalse("axNode.isChecked");
        node.accessibleNode.checked = "true";
        shouldBeEqualToString("node.accessibleNode.checked", "true");
        shouldBeTrue("axNode.isChecked");
    }

    function testCurrent() {
        debug("\nCurrent support");
        node = document.getElementById("current");
        axNode = accessibilityController.accessibleElementById("current");
        shouldBeEqualToString("axNode.stringAttributeValue('AXARIACurrent')", "false");
        node.accessibleNode.current = "page";
        shouldBeEqualToString("current.accessibleNode.current", "page");
        shouldBeEqualToString("axNode.stringAttributeValue('AXARIACurrent')", "page");
    }

    function testHasPopUp() {
        debug("\nHasPopUp support");
        node = document.getElementById("hasPopUp");
        axNode = accessibilityController.accessibleElementById("hasPopUp");
        shouldBeEqualToString("axNode.stringAttributeValue('AXHasPopUpValue')", "false");
        node.accessibleNode.hasPopUp = "dialog";
        shouldBeEqualToString("node.accessibleNode.hasPopUp", "dialog");
        shouldBeEqualToString("axNode.stringAttributeValue('AXHasPopUpValue')", "dialog");
    }

    function testInvalid() {
        debug("\nInvalid support");
        node = document.getElementById("invalid");
        axNode = accessibilityController.accessibleElementById("invalid");
        shouldBeEqualToString("axNode.stringAttributeValue('AXInvalid')", "false");
        node.accessibleNode.invalid = "grammar";
        shouldBeEqualToString("node.accessibleNode.invalid", "grammar");
        shouldBeEqualToString("axNode.stringAttributeValue('AXInvalid')", "grammar");
    }

    function testKeyShortcuts() {
        debug("\nKeyShortcuts support");
        node = document.getElementById("keyShortcuts");
        axNode = accessibilityController.accessibleElementById("keyShortcuts");
        node.accessibleNode.keyShortcuts = "Ctrl+F";
        shouldBeEqualToString("node.accessibleNode.keyShortcuts", "Ctrl+F");
        shouldBeEqualToString("axNode.stringAttributeValue('AXKeyShortcutsValue')", "Ctrl+F");
    }

    function testLive() {
        debug("\nLive support");
        node = document.getElementById("live");
        axNode = accessibilityController.accessibleElementById("live");
        shouldBeEqualToString("axNode.stringAttributeValue('AXARIALive')", "");
        node.accessibleNode.live = "polite";
        shouldBeEqualToString("node.accessibleNode.live", "polite");
        shouldBeEqualToString("axNode.stringAttributeValue('AXARIALive')", "polite");
    }

    function testOrientation() {
        debug("\nOrientation support");
        node = document.getElementById("orientation");
        axNode = accessibilityController.accessibleElementById("orientation");
        shouldBeEqualToString("axNode.orientation", "AXOrientation: AXHorizontalOrientation");
        node.accessibleNode.orientation = "vertical";
        shouldBeEqualToString("node.accessibleNode.orientation", "vertical");
        shouldBeEqualToString("axNode.orientation", "AXOrientation: AXVerticalOrientation");
    }

    function testPlaceholder() {
        debug("\nPlaceholder support");
        node = document.getElementById("placeholder");
        axNode = accessibilityController.accessibleElementById("placeholder");
        shouldBeEqualToString("axNode.stringAttributeValue('AXPlaceholderValue')", "");
        node.accessibleNode.placeholder = "test placeholder";
        shouldBeEqualToString("node.accessibleNode.placeholder", "test placeholder");
        shouldBeEqualToString("axNode.stringAttributeValue('AXPlaceholderValue')", "test placeholder");
    }

    function testPressed() {
        debug("\nPressed support");
        node = document.getElementById("pressed");
        axNode = accessibilityController.accessibleElementById("pressed");
        shouldBeFalse("axNode.boolAttributeValue('AXARIAPressedIsPresent')");
        node.accessibleNode.pressed = "true";
        shouldBeEqualToString("node.accessibleNode.pressed", "true");
        shouldBeTrue("axNode.boolAttributeValue('AXARIAPressedIsPresent')");
    }

    function testRelevant() {
        debug("\nRelevant support");
        node = document.getElementById("relevant");
        axNode = accessibilityController.accessibleElementById("relevant");
        shouldBeEqualToString("axNode.stringAttributeValue('AXARIARelevant')", "additions text");
        node.accessibleNode.relevant = "additions";
        shouldBeEqualToString("node.accessibleNode.relevant", "additions");
        shouldBeEqualToString("axNode.stringAttributeValue('AXARIARelevant')", "additions");
    }

    function testRoleDescription() {
        debug("\nRoleDescription support");
        node = document.getElementById("roleDescription");
        axNode = accessibilityController.accessibleElementById("roleDescription");
        shouldBeEqualToString("axNode.roleDescription", "AXRoleDescription: button");
        node.accessibleNode.roleDescription = "test button";
        shouldBeEqualToString("node.accessibleNode.roleDescription", "test button");
        shouldBeEqualToString("axNode.roleDescription", "AXRoleDescription: test button");
    }

    function testSort() {
        debug("\nSort support");
        node = document.getElementById("sort");
        axNode = accessibilityController.accessibleElementById("sort");
        shouldBeEqualToString("axNode.stringAttributeValue('AXSortDirection')", "AXUnknownSortDirection");
        node.accessibleNode.sort = "ascending";
        shouldBeEqualToString("node.accessibleNode.sort", "ascending");
        shouldBeEqualToString("axNode.stringAttributeValue('AXSortDirection')", "AXAscendingSortDirection");
    }

    function testValueText() {
        debug("\nValueText support");
        node = document.getElementById("valueText");
        axNode = accessibilityController.accessibleElementById("valueText");
        shouldBeEqualToString("axNode.valueDescription", "AXValueDescription: ");
        node.accessibleNode.valueText = "Five";
        shouldBeEqualToString("node.accessibleNode.valueText", "Five");
        shouldBeEqualToString("axNode.valueDescription", "AXValueDescription: Five");
    }

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
