LayoutTests/accessibility/mac/AOM-string-properties-expected.txt


This tests string type Accessibility Object Model properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Autocomplete support
PASS node.accessibleNode.autocomplete is "inline"
PASS axNode.stringAttributeValue('AXAutocompleteValue') is "inline"

Checked support
PASS axNode.isChecked is false
PASS node.accessibleNode.checked is "true"
PASS axNode.isChecked is true

Current support
PASS axNode.stringAttributeValue('AXARIACurrent') is "false"
PASS current.accessibleNode.current is "page"
PASS axNode.stringAttributeValue('AXARIACurrent') is "page"

HasPopUp support
PASS axNode.stringAttributeValue('AXHasPopUpValue') is "false"
PASS node.accessibleNode.hasPopUp is "dialog"
PASS axNode.stringAttributeValue('AXHasPopUpValue') is "dialog"

Invalid support
PASS axNode.stringAttributeValue('AXInvalid') is "false"
PASS node.accessibleNode.invalid is "grammar"
PASS axNode.stringAttributeValue('AXInvalid') is "grammar"

KeyShortcuts support
PASS node.accessibleNode.keyShortcuts is "Ctrl+F"
PASS axNode.stringAttributeValue('AXKeyShortcutsValue') is "Ctrl+F"

Live support
PASS axNode.stringAttributeValue('AXARIALive') is ""
PASS node.accessibleNode.live is "polite"
PASS axNode.stringAttributeValue('AXARIALive') is "polite"

Orientation support
PASS axNode.orientation is "AXOrientation: AXHorizontalOrientation"
PASS node.accessibleNode.orientation is "vertical"
PASS axNode.orientation is "AXOrientation: AXVerticalOrientation"

Placeholder support
PASS axNode.stringAttributeValue('AXPlaceholderValue') is ""
PASS node.accessibleNode.placeholder is "test placeholder"
PASS axNode.stringAttributeValue('AXPlaceholderValue') is "test placeholder"

Pressed support
PASS axNode.boolAttributeValue('AXARIAPressedIsPresent') is false
PASS node.accessibleNode.pressed is "true"
PASS axNode.boolAttributeValue('AXARIAPressedIsPresent') is true

Relevant support
PASS axNode.stringAttributeValue('AXARIARelevant') is "additions text"
PASS node.accessibleNode.relevant is "additions"
PASS axNode.stringAttributeValue('AXARIARelevant') is "additions"

RoleDescription support
PASS axNode.roleDescription is "AXRoleDescription: button"
PASS node.accessibleNode.roleDescription is "test button"
PASS axNode.roleDescription is "AXRoleDescription: test button"

Sort support
PASS axNode.stringAttributeValue('AXSortDirection') is "AXUnknownSortDirection"
PASS node.accessibleNode.sort is "ascending"
PASS axNode.stringAttributeValue('AXSortDirection') is "AXAscendingSortDirection"

ValueText support
PASS axNode.valueDescription is "AXValueDescription: "
PASS node.accessibleNode.valueText is "Five"
PASS axNode.valueDescription is "AXValueDescription: Five"
